requirements.txt
mysql-connector-python==8.0.33
cryptography==41.0.1
tkinter (built-in with Python)
